+++ fvm/tests/default_kernel/ops.rs
@@ -104,7 +104,7 @@ mod ipld {
                 .machine
                 .context()
                 .price_list
-                .on_block_create(block.len() as u32, 0)
+                .on_block_create(block.len(), 0)
                 .total();
             assert_eq!(
                 call_manager.gas_tracker.gas_used(),
@@ -181,7 +181,7 @@ mod ipld {
                 .machine
                 .context()
                 .price_list
-                .on_block_create(block.len() as u32, 0)
+                .on_block_create(block.len(), 0)
                 .total();
             let expected_link_price = call_manager
                 .machine
@@ -331,7 +331,7 @@ mod ipld {
         // assert gas
         {
             let price_list = call_manager.machine.context().price_list;
-            let expected_create_price = price_list.on_block_create(block.len() as u32, 0).total();
+            let expected_create_price = price_list.on_block_create(block.len(), 0).total();
             let expected_read_price = price_list.on_block_read(block.len()).total();
 
             assert_eq!(
@@ -397,7 +397,7 @@ mod ipld {
         // assert gas
         {
             let price_list = call_manager.machine.context().price_list;
-            let expected_create_price = price_list.on_block_create(block.len() as u32, 0).total();
+            let expected_create_price = price_list.on_block_create(block.len(), 0).total();
             let expected_stat_price = price_list.on_block_stat().total();
 
             assert_eq!(
